feat: Allow updating all equipment fields
Previously, updating equipment would fail if the 'id' field was present in the payload. This change enables the update operation to correctly handle all allowed fields for equipment, including the 'id' if it's inadvertently sent, by excluding it from the SQL SET clause. It also refactors the update logic to explicitly define and use a whitelist of updatable fields.